Move resources from sample app into core project
Created a resource bundle target that creates a resource bundle for the TimesSquare project. This allows applications to rely on the resource bundle for default look and feel.

Omitting the resource bundle from an application has no adverse side-effects (looks the same as it does before this patch)

Updated pod spec to generate the resource bundle